+++ Copilot/InlinePredictionManager.cs
@@ -13,6 +13,7 @@ using System.Text;
 using System.Text.RegularExpressions;
 using System.Threading;
 using System.Threading.Tasks;
+using System.Windows.Threading;
 using Constants = JeffPires.VisualChatGPTStudio.Utils.Constants;
 
 namespace JeffPires.VisualChatGPTStudio.Copilot
@@ -22,18 +23,28 @@ namespace JeffPires.VisualChatGPTStudio.Copilot
     /// </summary>
     internal class InlinePredictionManager
     {
+        private readonly OptionPageGridGeneral options;
         private readonly IWpfTextView view;
         private readonly ConcurrentDictionary<string, string> cache = new();
         private CancellationTokenSource cancellationTokenSource;
         private bool showingAutoComplete = false;
 
+        private readonly DispatcherTimer typingTimer;
+
         /// <summary>
         /// Initializes a new instance of the InlinePredictionManager class with the specified text view.
         /// </summary>
+        /// <param name="options">Extension's general options.</param>
         /// <param name="view">The IWpfTextView instance to be managed by the InlinePredictionManager.</param>
-        public InlinePredictionManager(IWpfTextView view)
+        public InlinePredictionManager(OptionPageGridGeneral options, IWpfTextView view)
         {
+            this.options = options;
             this.view = view;
+
+            typingTimer = new DispatcherTimer { Interval = TimeSpan.FromMilliseconds(options.CopilotSuggestionInterval) };
+            typingTimer.Tick += TypingTimer_Tick;
+
+            this.view.TextBuffer.Changed += TextBuffer_Changed;
         }
 
         /// <summary>
@@ -42,14 +53,12 @@ namespace JeffPires.VisualChatGPTStudio.Copilot
         /// It then sends a request to ChatGPT for a code prediction, processes the prediction, and displays
         /// the autocomplete suggestion in the editor.
         /// </summary>
-        public async Task OnEnterPressed(OptionPageGridGeneral options)
+        public async Task ShowAutocompleteAsync()
         {
             const string MARKER = "\n **AUTOCOMPLETE HERE** \n";
 
             try
             {
-                await VS.StatusBar.ShowProgressAsync(Constants.MESSAGE_WAITING_COPILOT, 1, 2);
-
                 cancellationTokenSource?.Cancel();
                 cancellationTokenSource = new();
 
@@ -103,10 +112,6 @@ namespace JeffPires.VisualChatGPTStudio.Copilot
             {
                 showingAutoComplete = false;
                 Logger.Log(ex);
-            }
-            finally
-            {
-                await VS.StatusBar.ShowProgressAsync(Constants.MESSAGE_WAITING_COPILOT, 2, 2);
             }
         }
 
@@ -274,6 +279,36 @@ namespace JeffPires.VisualChatGPTStudio.Copilot
                 {
                     cache.TryRemove(key, out _);
                 }
+            }
+        }
+
+        /// <summary>
+        /// Handles the event triggered when the text buffer changes.
+        /// </summary>
+        private void TextBuffer_Changed(object sender, TextContentChangedEventArgs e)
+        {
+            typingTimer.Stop();
+            typingTimer.Start();
+        }
+
+        /// <summary>
+        /// Handles the tick event of the typing timer, stopping the timer and triggering a suggestion asynchronously.
+        /// </summary>
+        private async void TypingTimer_Tick(object sender, EventArgs e)
+        {
+            typingTimer.Stop();
+
+            await TriggerSuggestionAsync();
+        }
+
+        /// <summary>
+        /// Triggers the suggestion mechanism asynchronously if it is not already active.
+        /// </summary>
+        private async Task TriggerSuggestionAsync()
+        {
+            if (!showingAutoComplete)
+            {
+                await ShowAutocompleteAsync();
             }
         }
     }